% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_to_ret.R
\name{adapt_to_ret}
\alias{adapt_to_ret}
\title{Auxiliary function for enabling correct conversion from R objects to python and GAMS}
\usage{
adapt_to_ret(set_par)
}
\arguments{
\item{set_par}{a set or a parameter to be loaded in ixmp}
}
\description{
Auxiliary function for enabling correct conversion from R objects to python and GAMS
}
